// src/app/models/user-management/memberInfoData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

import { MemberAddressData } from './memberAddressData';

export interface MemberInfoData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** สถานะการศึกษา/การทำงานปัจจุบัน */
    currentJob?: number;
    /** facebook_token */
    facebookToken?: string;
    first?: number;
    /** gmail_token */
    gmailToken?: string;
    /** วุฒิการศึกษาสูงสุด (ที่ได้รับ) */
    maxEducation?: number;
    /** วัน/เดือน/ปีเกิด */
    memberBirthdate?: string;
    /** เลขประจำตัวประชาชน/หมายเลขพาสปอร์ต */
    memberCardno?: string;
    /** รูปแบบบัญชีที่ใช้สมัคร */
    memberChannel?: number;
    /** สัญชาติ */
    countryId?: number;
    /** ประเภทสัญชาติผู้สมัคร */
    memberCountryType?: number;
    /** อีเมล */
    memberEmail?: string;
    /** ชื่อ (อังกฤษ) */
    memberFirstnameEn?: string;
    /** ชื่อ (ไทย) */
    memberFirstnameTh?: string;
    /** เพศ */
    memberGender?: number;
    /** PK */
    memberId?: number;
    /** นามสกุล (อังกฤษ) */
    memberLastnameEn?: string;
    /** นามสกุล (ไทย) */
    memberLastnameTh?: string;
    /** โทรศัพท์ */
    memberPhoneno?: string;

    /** รูปผู้ลงทะเบียนเรียน */
    filename?: string;
    prefix?: string;
    module?: number;

    mode?: string;
    /** อ้างอิงรหัส moodle user id (PK) */
    moodleUserId?: number;
    /** คำนำหน้า */
    prefixnameId?: number;
    /** สถานะการสมัครรับข่าวสาร */
    receivedNews?: boolean;
    rowNum?: number;
    size?: number;
    channelTh?: string;
    channelEn?: string;
    memberAdderss?: MemberAddressData;
    lastLoginDatetime?: string;
    lastLogin?: string;

    username?: string;

    memberNo?: string;
}